build: moved main to cmd
