+++ .gitignore
@@ -11,19 +11,12 @@
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
 
-# Dependency directories (remove the comment below to include it)
-# vendor/
-
-# Do not commit the executable
+# Do not commit executables
 logo-ls
 build/
 
 # distribution directory
 dist
-
-# env stuff
-*.env
-.gh_token
 
 # test related data
 testdata/dirEnv*
